Move BookList.js and BookForm.js to containers directory from components directory
