+++ src/containers/BooksList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import Book from './Book';
+import Book from '../components/Book';
 
 const BooksList = ({ books }) => {
   const booksArr = books.map(book => (
